services/ai-python/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ai-advisor"
version = "0.1.0"
description = "LangChain-powered RAG advisor for the kitchen configurator"
readme = "README.md"
authors = [{ name = "Kitchen Configurator", email = "oss@kitchen.example" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "langchain>=0.2.1,<0.3.0",
  "langchain-text-splitters>=0.2.1,<0.3.0",
  "langchain-openai>=0.1.7,<0.2.0",
  "pgvector>=0.2.5,<0.3.0",
  "psycopg[binary,pool]>=3.1.20,<4.0.0",
  "sqlalchemy>=2.0.34,<3.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.2.1,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "numpy>=1.26.4,<3.0.0",
  "faiss-cpu>=1.7.4",
  "structlog>=24.1.0,<25.0.0",
  "tenacity>=8.2.3,<9.0.0",
  "httpx>=0.27.0,<0.28.0",
  "opentelemetry-sdk>=1.26.0,<2.0.0",
  "opentelemetry-exporter-otlp>=1.26.0,<2.0.0",
  "opentelemetry-instrumentation-fastapi>=0.52.0,<0.53.0",
  "opentelemetry-instrumentation-requests>=0.52.0,<0.53.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2,<9.0.0",
  "pytest-asyncio>=0.23.8,<0.24.0",
  "ruff>=0.6.0,<0.7.0",
  "mypy>=1.11.0,<2.0.0"
]

[project.scripts]
ai-advisor-api = "ai_advisor.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_advisor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
